Optimized menu and go back integrated.

+++ gpg-messages-manager.sh
@@ -98,20 +98,26 @@ function showMenu() {
 
 function askOption() {
 
-	choice=$(echo -e "[0] Show keyring
-[1] Generate a new key pair
-[2] Import a public key
-[3] Export a public key
-[4] Remove a public key
-[5] Encrypt your address
-[6] Encrypt a message
-[7] Decrypt a message
-[8] Sign a file
-[9] Verify a file
-[!] Exit" |
+	options=(
+		"Show keyring"
+		"Generate a new key pair"
+		"Import a public key"
+		"Export a public key"
+		"Remove a public key"
+		"Encrypt your addresse"
+		"Encrypt a message"
+		"Decrypt a message"
+		"Sign a file"
+		"Verify a file"
+		"Exit"
+	)
+
+	choice=$(for (( i=0; i<${#options[*]}; i++ )); do
+		echo "[${i}] ${options[${i}]}"
+	done |
 	$fzf |
 	cut -d ' ' -f1 |
-	grep -oE '[0-9]+'
+	grep -oE '([0-9])'
 	)
 
 	case $choice in
@@ -136,6 +142,7 @@ function askOption() {
 		9)
 			verifyFile;;
 	esac
+	clear
 }
 
 function showKeys() {
@@ -165,12 +172,15 @@ function selectKey() {
 }
 
 function genKey() {
+
 	gpgVersion=$(gpg --version | sed 1q | cut -d ' ' -f3)
 
 	if [ $(echo -e "2.2.17\n${gpgVersion}" | sort -V | head -n1) == "2.2.17" ]; then
+
 		# GPG version is greather than 2.2.17
 		gpg --full-generate-key
 	else
+
 		# GPG version is lower than 2.2.17
 		gpg --gen-key
 	fi
@@ -179,29 +189,39 @@ function genKey() {
 }
 
 function addPubKey() {
-	cd $gpgDir
-	echo "Please select a public key : "
-	public=$(ls pub/ | head | $fzf)
-	gpg --batch --yes --import pub/$public
 
-	if [ $? -eq 0 ]; then
-		echo -e "${green}\nPublic key $public added !${end}"
-	else
-		echo -e "\n${red}Warning: An error occured. Please refer to the gpg error log.${end}"
+	cd $gpgDir
+	echo "Please select a public key to import : "
+
+	public=$(
+		ls pub/ |
+		head |
+		$fzf
+	)
+
+	if [ ! -z $public ]; then
+		gpg --batch --yes --import pub/$public
+		if [ $? -eq 0 ]; then
+			echo -e "${green}\nPublic key $public added !${end}"
+		else
+			echo -e "\n${red}Warning: An error occured. Please refer to the gpg error log.${end}"
+		fi
 	fi
 
 	toMenu
 }
 
 function delPubKey() {
-	echo "Please select a public key : "
+	echo "Please select a public key to delete : "
 	public=$(selectKey)
 	gpg --batch --yes --delete-key $public
 
-	if [ $? -eq 0 ]; then
-		echo -e "${green}Public key $public deleted !${end}"
-	else
-		echo -e "\n${red}Warning: An error occured. Please refer to the gpg error log..${end}"
+	if [ ! -z $public ]; then
+		if [ $? -eq 0 ]; then
+			echo -e "${green}Public key $public deleted !${end}"
+		else
+			echo -e "\n${red}Warning: An error occured. Please refer to the gpg error log..${end}"
+		fi
 	fi
 
 	toMenu
@@ -212,33 +232,37 @@ function encryptMsg() {
 	echo "Please select a recipient : "
 	recipient=$(selectKey)
 
-	if [ "$1" == "adr" ]; then
-		gpg --trust-model always --armor --encrypt --recipient $recipient txt/adr.txt
+	if [ ! -z $recipient ]; then
 
-		if [ $? -eq 0 ]; then
-			recipientFormat=$(echo -n 'adr_for_'; echo $recipient | sed 's/@.\+$//')
-			mv txt/adr.txt.asc asc/$recipientFormat.asc
+		if [ "$1" == "adr" ]; then
+			gpg --trust-model always --armor --encrypt --recipient $recipient txt/adr.txt
 
-			echo -e "\n${green}Address encrypted for ${bs}${recipient}${be}\nto $gpgDir/asc/${bs}$recipientFormat.asc${be} !${end}"
+			if [ $? -eq 0 ]; then
+				recipientFormat=$(echo -n 'adr_for_'; echo $recipient | sed 's/@.\+$//')
+				mv txt/adr.txt.asc asc/$recipientFormat.asc
+
+				echo -e "\n${green}Address encrypted for ${bs}${recipient}${be}\nto $gpgDir/asc/${bs}$recipientFormat.asc${be} !${end}"
+			else
+				echo -e "\n${red}Warning: An error occured. Please refer to the gpg error log..${end}"
+			fi
+
 		else
-			echo -e "\n${red}Warning: An error occured. Please refer to the gpg error log..${end}"
+			$editor txt/tmp.txt
+			gpg --trust-model always --armor --encrypt --recipient $recipient txt/tmp.txt
+
+			if [ $? -eq 0 ]; then
+				recipientFormat=$(echo -n 'msg_for_'; echo -n $recipient | sed 's/@.\+$/_/'; date +%Y%m%d_%H%M%S)
+				mv txt/tmp.txt.asc asc/$recipientFormat.asc
+				rm txt/tmp.txt
+
+				echo -e "\n${green}Message encrypted for ${bs}${recipient}${be}\nto $gpgDir/asc/${bs}$recipientFormat.asc${be} !${end}"
+
+				# Would you like to sign it ?
+			else
+				echo -e "\n${red}Warning: An error occured. Please refer to the gpg error log..${end}"
+			fi
 		fi
 
-	else
-		$editor txt/tmp.txt
-		gpg --trust-model always --armor --encrypt --recipient $recipient txt/tmp.txt
-
-		if [ $? -eq 0 ]; then
-			recipientFormat=$(echo -n 'msg_for_'; echo -n $recipient | sed 's/@.\+$/_/'; date +%Y%m%d_%H%M%S)
-			mv txt/tmp.txt.asc asc/$recipientFormat.asc
-			rm txt/tmp.txt
-
-			echo -e "\n${green}Message encrypted for ${bs}${recipient}${be}\nto $gpgDir/asc/${bs}$recipientFormat.asc${be} !${end}"
-
-			# Would you like to sign it ?
-		else
-			echo -e "\n${red}Warning: An error occured. Please refer to the gpg error log..${end}"
-		fi
 	fi
 
 	echo 'RELOADAGENT' | gpg-connect-agent
@@ -248,25 +272,30 @@ function encryptMsg() {
 function decryptMsg() {
 	cd $gpgDir
 	messages=$(ls asc/ | grep '^msg' | head -n99)
-	echo "Please select a message : "
+	echo "Please select a message to decrypt : "
 	toDecryptMsg=$(echo -e "$messages" | $fzf)
-	gpg --output txt/$toDecryptMsg.txt --decrypt asc/$toDecryptMsg
 
-	if [ $? -eq 0 ]; then
-		echo -e "\n${green}Message successfully decrypted to $gpgDir/txt/${bs}$toDecryptMsg.txt${be} !${end}"
-		echo -e "\nWould you like to see the message now ?\n[Y]es / [N]o\n"
-		read -p ">>> " answerMsg
+	if [ ! -z $toDecryptMsg ]; then
 
-		while [[ ! $answerMsg == "Y" && ! $answerMsg == "y" && ! $answerMsg == "N" && ! $answerMsg == "n" ]]; do
-			echo -e "\nPlease choose between Y or N !"
+		gpg --output txt/$toDecryptMsg.txt --decrypt asc/$toDecryptMsg
+
+		if [ $? -eq 0 ]; then
+			echo -e "\n${green}Message successfully decrypted to $gpgDir/txt/${bs}$toDecryptMsg.txt${be} !${end}"
+			echo -e "\nWould you like to see the message now ?\n[Y]es / [N]o\n"
 			read -p ">>> " answerMsg
-		done
 
-		if [[ $answerMsg == "Y" || $answerMsg == "y" ]]; then
-			$editor txt/$toDecryptMsg.txt
+			while [[ ! $answerMsg == "Y" && ! $answerMsg == "y" && ! $answerMsg == "N" && ! $answerMsg == "n" ]]; do
+				echo -e "\nPlease choose between Y or N !"
+				read -p ">>> " answerMsg
+			done
+
+			if [[ $answerMsg == "Y" || $answerMsg == "y" ]]; then
+				$editor txt/$toDecryptMsg.txt
+			fi
+		else
+			echo -e "\n${red}Warning: An error occured. Please refer to the gpg error log..${end}"
 		fi
-	else
-		echo -e "\n${red}Warning: An error occured. Please refer to the gpg error log..${end}"
+
 	fi
 
 	echo 'RELOADAGENT' | gpg-connect-agent
@@ -275,15 +304,20 @@ function decryptMsg() {
 
 function exportPubKey() {
 	cd $gpgDir
-	echo -e "\nPlease select a public key : "
+	echo -e "\nPlease select a public key to export : "
 	public=$(selectKey)
-	pubFormat=$(echo -n $public | sed 's/@.\+$/_/' | sed 's/\./_/g'; echo 'pub.asc')
-	gpg -ao pub/${pubFormat} --export ${public}
 
-	if [ $? -eq 0 ]; then
-		echo -e "\n${green}Public key successfully exported to $gpgDir/pub/${bs}${pubFormat}${be} !${end}"
-	else
-		echo -e "\n${red}Warning: An error occured. Please refer to the gpg error log.${end}"
+	if [ ! -z $public ]; then
+
+		pubFormat=$(echo -n $public | sed 's/@.\+$/_/' | sed 's/\./_/g'; echo 'pub.asc')
+		gpg -ao pub/${pubFormat} --export ${public}
+
+		if [ $? -eq 0 ]; then
+			echo -e "\n${green}Public key successfully exported to $gpgDir/pub/${bs}${pubFormat}${be} !${end}"
+		else
+			echo -e "\n${red}Warning: An error occured. Please refer to the gpg error log.${end}"
+		fi
+
 	fi
 
 	echo RELOADAGENT | gpg-connect-agent
@@ -292,14 +326,19 @@ function exportPubKey() {
 
 function signFile() {
 	cd $gpgDir
-	echo -e "\nPlease select a file : "
+	echo -e "\nPlease select a file to sign : "
 	file=$(ls asc/ txt/ | sed '/.*\/:/d' | sed '/^$/d' | head -n99 | $fzf)
-	gpg --armor --output sig/${file}.sig --detach-sig $(find | grep $file | tail -n1)
 
-	if [ $? -eq 0 ]; then
-		echo -e "\n${green}File successfully signed to $gpgDir/sig/${bs}${file}.sig${be} !${end}"
-	else
-		echo -e "\n${red}Warning: An error occured. Please refer to the gpg error log.${end}"
+	if [ ! -z $file ]; then
+
+		gpg --armor --output sig/${file}.sig --detach-sig $(find | grep $file | tail -n1)
+
+		if [ $? -eq 0 ]; then
+			echo -e "\n${green}File successfully signed to $gpgDir/sig/${bs}${file}.sig${be} !${end}"
+		else
+			echo -e "\n${red}Warning: An error occured. Please refer to the gpg error log.${end}"
+		fi
+
 	fi
 
 	echo RELOADAGENT | gpg-connect-agent
@@ -308,13 +347,18 @@ function signFile() {
 
 function verifyFile() {
 	cd $gpgDir
-	echo -e "\nPlease select a file : "
+	echo -e "\nPlease select a file to verify : "
 	signature=$(ls sig/ | $fzf)
-	fileSign=$(find asc/ txt/ | grep $(echo $signature | sed 's/\.sig$//'))
-	gpg --verify sig/${signature} ${fileSign}
 
-	if [ $? -ne 0 ]; then
-		exit 1
+	if [ ! -z $signature ]; then
+
+		fileSign=$(find asc/ txt/ | grep $(echo $signature | sed 's/\.sig$//'))
+		gpg --verify sig/${signature} ${fileSign}
+
+		if [ $? -ne 0 ]; then
+			exit 1
+		fi
+
 	fi
 
 	echo RELOADAGENT | gpg-connect-agent
